Fix linting errors and split up test files
