+++ lib/is-valid-selector.js
@@ -21,35 +21,37 @@ module.exports = isValidSelector;
  * - A function that accepts a component name and returns a regexp.
  *   The returned regexp will be used to test all selector sequences.
  * - An object that has the following properties:
- *   - initial: A function that accepts a component name and returns a regexp.
- *     This regexp will be used to test selector sequences at the beginning of the
- *     selector. If no chainable property is included, this function will also be
- *     used to test chained sequences.
- *   - chainable: A function that accepts a component name and returns a regexp.
- *     This regexp will be used to test selector sequences that are chained
- *     (not the first one), if they do not already match the initial pattern.
+ *   - initial: A function that accepts a component name and returns
+ *   	 a regexp. This regexp will be used to test selector sequences
+ *   	 at the beginning of the selector. If no chainable property is
+ *   	 included, this function will also be used to test chained
+ *   	 sequences.
+ *   - chainable: A function that accepts a component name and returns
+ *   	 a regexp. This regexp will be used to test selector sequences
+ *   	 that are chained (not the first one), if they do not already
+ *   	 match the initial pattern.
  */
 
 /**
  * @param {String} selector - The selector to test
  * @param {String} componentName - The component's name
- * @param {Boolean} strict - Whether to use strict mode
- * @param {String|SelectorPattern} [patternDef="suit"] - Either a string
+ * @param {Boolean} isStrictMode - Whether to use strict mode
+ * @param {String|SelectorPattern} [pattern="suit"] - Either a string
  *   identifying a pre-defined SelectorPattern to use, or a custom
  *   SelectorPattern, as described above
  * @returns {Boolean}
  */
-function isValidSelector(selector, componentName, isStrictMode, selectorPattern) {
+function isValidSelector(selector, componentName, isStrictMode, pattern) {
 
   // Don't bother with :root
   if (selector === ':root') { return true; }
 
-  var initialPattern = (selectorPattern.initial)
-    ? selectorPattern.initial(componentName)
-    : selectorPattern(componentName);
-  var chainablePattern = (selectorPattern.chainable)
-    ? selectorPattern.chainable(componentName)
-    : initialPattern;
+  var initialPattern = (pattern.initial) ?
+    pattern.initial(componentName) :
+    pattern(componentName);
+  var chainablePattern = (pattern.chainable) ?
+    pattern.chainable(componentName) :
+    initialPattern;
   var simpleSelectors = listSimpleSelectors(selector);
 
   // Error if an acceptable initialPattern does not begin the selector
@@ -59,8 +61,8 @@ function isValidSelector(selector, componentName, isStrictMode, selectorPattern)
   // chainablePattern
   if (isStrictMode) {
     return simpleSelectors.slice(1).every(function (chainedSelector) {
-      return initialPattern.test(chainedSelector)
-        || chainablePattern.test(chainedSelector);
+      return initialPattern.test(chainedSelector) ||
+      chainablePattern.test(chainedSelector);
     });
   }
