+++ lib/preset-patterns.js
@@ -37,6 +37,6 @@ function bemSelector(block) {
   var modifier = '(?:(?:_' + WORD + '){1,2})';
   return new RegExp(
     '^\\.' + block + modifier + '?$' +
-    '|^\\.' + block + element + modifier +'?$'
+    '|^\\.' + block + element + modifier + '?$'
   );
 }
