+++ lib/validate-selectors.js
@@ -4,7 +4,7 @@
  * Module dependencies
  */
 
-var isValidSelectorInComponent = require('./is-valid-selector');
+var isValid = require('./is-valid-selector');
 
 /**
  * Module exports
@@ -16,7 +16,7 @@ module.exports = validateSelectors;
  * @param {Object} styles
  * @param {String} componentName
  */
-function validateSelectors(styles, componentName, strict, selectorPattern) {
+function validateSelectors(styles, componentName, strict, pattern) {
   styles.eachRule(function (rule) {
     if (rule.parent && rule.parent.name == 'keyframes') {
       return;
@@ -25,7 +25,7 @@ function validateSelectors(styles, componentName, strict, selectorPattern) {
 
     selectors.forEach(function (selector) {
       // selectors must start with the componentName class, or be `:root`
-      if (!isValidSelectorInComponent(selector, componentName, strict, selectorPattern)) {
+      if (!isValid(selector, componentName, strict, pattern)) {
         throw rule.error('Invalid selector "' + selector + '". ');
       }
     });
